# Optimization: tiled versions of JTDAJ