--- a/mujoco_warp/_src/io.py
+++ b/mujoco_warp/_src/io.py
@@ -935,10 +935,15 @@ def _mujoco_octree_to_warp_volume(
   return volume_ids_array, volumes_tuple, oct_aabb_array
 
 
-def get_padded_sizes(nv: int, njmax: int, nworld: int, is_sparse: bool, tile_size):
+def get_padded_sizes(nv: int, njmax: int, nworld: int, is_sparse: bool):
   # if dense - we just pad to the next multiple of 4 for nv, to get the fast load path.
   #            we pad to the next multiple of tile_size for njmax to avoid out of bounds accesses.
   # if sparse - we pad to the next multiple of tile_size for njmax, and nv.
+
+  if is_sparse:
+    tile_size = 16
+  else:
+    tile_size = 32
 
   def round_up(x, multiple):
     return ((x + multiple - 1) // multiple) * multiple
@@ -1006,9 +1011,7 @@ def make_data(mjm: mujoco.MjModel, nworld: int = 1, nconmax: int = -1, njmax: in
   nsensorcontact = np.sum(mjm.sensor_type == mujoco.mjtSensor.mjSENS_CONTACT)
   nrangefinder = sum(mjm.sensor_type == mujoco.mjtSensor.mjSENS_RANGEFINDER)
 
-  TILE_SIZE = 16
-
-  J_padded_size, h_padded_size, d_padded_size, state_padded_size = get_padded_sizes(mjm.nv, njmax, nworld, mujoco.mj_isSparse(mjm), TILE_SIZE)
+  J_padded_size, h_padded_size, d_padded_size, state_padded_size = get_padded_sizes(mjm.nv, njmax, nworld, mujoco.mj_isSparse(mjm))
 
   return types.Data(
     nworld=nworld,
@@ -1328,9 +1331,7 @@ def put_data(
   ne_jnt = int(np.sum((mjm.eq_type == mujoco.mjtEq.mjEQ_JOINT) & mjd.eq_active))
   ne_ten = int(np.sum((mjm.eq_type == mujoco.mjtEq.mjEQ_TENDON) & mjd.eq_active))
 
-  TILE_SIZE = 16
-
-  J_padded_size, h_padded_size, d_padded_size, state_padded_size = get_padded_sizes(mjm.nv, njmax, nworld, mujoco.mj_isSparse(mjm), TILE_SIZE)
+  J_padded_size, h_padded_size, d_padded_size, state_padded_size = get_padded_sizes(mjm.nv, njmax, nworld, mujoco.mj_isSparse(mjm))
 
   efc_type_fill = np.zeros((nworld, njmax))
   efc_id_fill = np.zeros((nworld, njmax))