--- a/mujoco_warp/_src/solver.py
+++ b/mujoco_warp/_src/solver.py
@@ -1459,8 +1459,8 @@ def compute_jtdj_tiled_kernel(
         if k > nefc:
             break
 
-        J_ki = wp.tile_load(efc_J_in[worldid], shape=(TILE_SIZE, TILE_SIZE), offset=(k, i))
-        J_kj = wp.tile_load(efc_J_in[worldid], shape=(TILE_SIZE, TILE_SIZE), offset=(k, j))
+        J_ki = wp.tile_load(efc_J_in[worldid], shape=(TILE_SIZE, TILE_SIZE), offset=(k, i * TILE_SIZE))
+        J_kj = wp.tile_load(efc_J_in[worldid], shape=(TILE_SIZE, TILE_SIZE), offset=(k, j * TILE_SIZE))
         D_k = wp.tile_load(efc_D_in[worldid], shape=TILE_SIZE, offset=k)
         state = wp.tile_load(efc_state_in[worldid], shape=TILE_SIZE, offset=k)
 
@@ -1479,7 +1479,9 @@ def compute_jtdj_tiled_kernel(
         J_ki = wp.tile_map(wp.mul, wp.tile_transpose(J_ki), wp.tile_broadcast(D_k, shape=(TILE_SIZE, TILE_SIZE)))
         sum_val += wp.tile_matmul(J_ki, J_kj)
 
-    wp.tile_store(efc_h_out[worldid], sum_val, offset=(i, j))
+    wp.tile_store(efc_h_out[worldid], sum_val, offset=(i * TILE_SIZE, j * TILE_SIZE))
+    #ones = wp.tile_ones(shape=(TILE_SIZE, TILE_SIZE), dtype=wp.float32)
+    #wp.tile_store(efc_h_out[worldid], ones, offset=(i * TILE_SIZE, j * TILE_SIZE))
 
 
 @wp.kernel
@@ -1743,6 +1745,8 @@ def _update_gradient(m: types.Model, d: types.Data):
     # h = qM + (efc_J.T * efc_D * active) @ efc_J
     lower_triangle_dim = int(m.nv * (m.nv + 1) / 2)
     if m.opt.is_sparse:
+      # Run both kernels for comparison
+      # First, run the original sparse kernel
       if 0:
         wp.launch(
           update_gradient_JTDAJ_sparse,
@@ -1758,8 +1762,9 @@ def _update_gradient(m: types.Model, d: types.Data):
           outputs=[d.efc.h],
         )
       else:
+        # Create temp array and run tiled kernel
+        #temp_efc_h = wp.zeros(shape=(d.nworld, m.nv, m.nv), dtype=wp.float32)
         num_blocks_ceil = ceil(m.nv / TILE_SIZE)
-        #lower_triangle_dim_blocked = int(num_blocks_ceil * (num_blocks_ceil + 1) / 2)
         wp.launch(
           compute_jtdj_tiled_kernel,
           dim=(d.nworld, num_blocks_ceil, num_blocks_ceil, BLOCK_DIM),
@@ -1774,6 +1779,26 @@ def _update_gradient(m: types.Model, d: types.Data):
           outputs=[d.efc.h],
           block_dim=BLOCK_DIM,
         )
+      
+      # Compare results using numpy
+      #wp.synchronize()
+      #sparse_result = d.efc.h.numpy()
+      #tiled_result = temp_efc_h.numpy()
+      
+      # Compare lower triangular elements (including diagonal)
+      #for worldid in range(d.nworld):
+      #  if not d.efc.done.numpy()[worldid]:
+      #    for i in range(m.nv):
+      #      for j in range(i + 1):  # Only lower triangular + diagonal (j <= i)
+      #        sparse_val = sparse_result[worldid, i, j]
+      #        tiled_val = tiled_result[worldid, i, j]
+
+      #        #print(f"World {worldid}, pos({i},{j}): sparse={sparse_val:.6f}, tiled={tiled_val:.6f}")
+      #        
+      #        if abs(sparse_val - tiled_val) > 1e-4:
+      #          print(f"World {worldid}, pos({i},{j}): sparse={sparse_val:.6f}, tiled={tiled_val:.6f}")
+      
+      
       wp.launch(
         update_gradient_set_h_qM_lower_sparse,
         dim=(d.nworld, m.qM_fullm_i.size),