Significant speedups, but still WIP as I need to verify the benchmarks everywhere. Thought I'd share a preview :)