@adenzler-nvidia 2025-09-03
some numbers (total steps per second) using parallel linesearch

| Benchmark | main | this branch | speedup |
|-----------------|---------|----------------| ----------- |
| humanoid | 3,583,055 |3,753,086| 1.05|
| n_humanoid |358,086 | 485,103 | 1.35 |
| aloha pot |1,450,175 |1,383,222| 0.95 |
| apollo | 2,654,118 | 2,717,492 | 1.02 |
| kitchen G1 (4k worlds) | 17,338 | 21,103 | 1.21 | 
| kitchen anymal (4k worlds) | 51,271 | 67,883 | 1.32 |
| franka | 13,997,320 | 13,045,987 | 0.93 |

Will look into the regressions now. Will probably have to think about the dense version some more.