@adenzler-nvidia 2025-09-05
This is ready for review, still looking at the test failure though, I can'r reproduce it on my system.

I think something is off with the aloha pot such that the numbers are weird, so I wouldn't trust that one too much. The franka numbers are kind of expected as that benchmark has very few actual constraints.

The numbers look great for sparse systems with big constraint counts. For dense systems, the numbers don't look that impressive  but I can see clear gains on the kernel time itself. The kernel is much less dominant which means we're not seeing a lot of the effects in total, but it's still one of the expensive ones and I think this is just a stepping stone.