@adenzler-nvidia 2025-10-03
updated numbers: 

| Benchmark | main | this branch | 
|-----------------|---------|----------------| 
| humanoid | 3,983,425 | 4,074,985 | 
| n_humanoid | 365,992 | 509,190 | 
| aloha pot | 2,143,688 | 2,146,836 | 
| apollo | 2,752,688 | 2,803,341 | 
| kitchen G1 (4k worlds) | 21,798 | 29,940 | 
| franka | 14,411,979 | 13,957,684 |

For the ones using dense the speedups are not massive but it's clearly visible in the kernel runtimes. There is further potential now in fusing in the cholesky solve as well but I will defer that to follow-up PR as it once again requires tuning a few more thresholds, block-sizes.